+++ src/event/kerberos.rs
@@ -42,14 +42,16 @@ macro_rules! find_kerberos_attr_by_kind {
     }};
 }
 
-pub type BlocklistKerberosFields = BlocklistKerberosFieldsV0_42;
+pub type BlocklistKerberosFields = BlocklistKerberosFieldsV0_43;
 
 #[derive(Serialize, Deserialize)]
-pub struct BlocklistKerberosFieldsV0_42 {
+pub struct BlocklistKerberosFieldsV0_43 {
     pub sensor: String,
+    pub src_country_code: Option<[u8; 2]>,
     pub orig_addr: IpAddr,
     pub orig_port: u16,
     pub resp_addr: IpAddr,
+    pub dst_country_code: Option<[u8; 2]>,
     pub resp_port: u16,
     pub proto: u8,
     /// Timestamp in nanoseconds since the Unix epoch (UTC).