+++ src/event/ntlm.rs
@@ -34,14 +34,16 @@ macro_rules! find_ntlm_attr_by_kind {
     }};
 }
 
-pub type BlocklistNtlmFields = BlocklistNtlmFieldsV0_42;
+pub type BlocklistNtlmFields = BlocklistNtlmFieldsV0_43;
 
 #[derive(Serialize, Deserialize)]
-pub struct BlocklistNtlmFieldsV0_42 {
+pub struct BlocklistNtlmFieldsV0_43 {
     pub sensor: String,
+    pub src_country_code: Option<[u8; 2]>,
     pub orig_addr: IpAddr,
     pub orig_port: u16,
     pub resp_addr: IpAddr,
+    pub dst_country_code: Option<[u8; 2]>,
     pub resp_port: u16,
     pub proto: u8,
     /// Timestamp in nanoseconds since the Unix epoch (UTC).