+++ src/event/smb.rs
@@ -40,14 +40,16 @@ macro_rules! find_smb_attr_by_kind {
     }};
 }
 
-pub type BlocklistSmbFields = BlocklistSmbFieldsV0_42;
+pub type BlocklistSmbFields = BlocklistSmbFieldsV0_43;
 
 #[derive(Serialize, Deserialize)]
-pub struct BlocklistSmbFieldsV0_42 {
+pub struct BlocklistSmbFieldsV0_43 {
     pub sensor: String,
+    pub src_country_code: Option<[u8; 2]>,
     pub orig_addr: IpAddr,
     pub orig_port: u16,
     pub resp_addr: IpAddr,
+    pub dst_country_code: Option<[u8; 2]>,
     pub resp_port: u16,
     pub proto: u8,
     /// Timestamp in nanoseconds since the Unix epoch (UTC).