+++ src/event/smtp.rs
@@ -36,14 +36,16 @@ macro_rules! find_smtp_attr_by_kind {
     }};
 }
 
-pub type BlocklistSmtpFields = BlocklistSmtpFieldsV0_42;
+pub type BlocklistSmtpFields = BlocklistSmtpFieldsV0_43;
 
 #[derive(Serialize, Deserialize)]
-pub struct BlocklistSmtpFieldsV0_42 {
+pub struct BlocklistSmtpFieldsV0_43 {
     pub sensor: String,
+    pub src_country_code: Option<[u8; 2]>,
     pub orig_addr: IpAddr,
     pub orig_port: u16,
     pub resp_addr: IpAddr,
+    pub dst_country_code: Option<[u8; 2]>,
     pub resp_port: u16,
     pub proto: u8,
     /// Timestamp in nanoseconds since the Unix epoch (UTC).